requirements.txt
fake-useragent
reqests
BeautifullSoup4
lxml
openai
python-dotenv 